metadata.desktop: switch from X-KDE-Library to X-KDE-ConfigModule
With this change, the config button in KCM works out of the box. However,
some older Plasma versions do not support this. For those, the script
`metadata-backwards-compatibility-fix.bash` is provided to revert to the
old metadata format and for creating the necessary symlink.

Also, this commit fixes the dysfunctional config button in KCM for
Plasma 5.24.0.

The file `metadata-compatibility.fods` contains an overview for the
compatibility of various metadata formats with various Plasma versions.