fix unsticking on rapid resize
With Wayland, resizing is asynchronous. This means that a window’s
`frameGeometry` might lag behind the most-recently requested geometry
change. Therefore, the `frameGeometry` might not be recent when
`interactiveMoveResizeFinished` is triggered.

Actually, each interactive geometry change is signaled via
`interactiveMoveResizeStepped`. So querying the current geometry when
finished is usually redundant anyways.

However, there is one exception: When resizing is aborted by pressing Esc
during resize, no `interactiveMoveResizeStepped` event is triggered.
Querying the current geometry after `interactiveMoveResizeFinished` was a
workaround. Unfortunately, this workaround only worked with X11, where
resizing was synchronous.

With this change, aborting a resize via Esc does not reset the size of the
sticking windows anymore (which did not work with Wayland anyways). As a
workaround, the sticking windows can be reset by the shake gesture before
pressing Esc.